weave: Move parser state into a passed value
To allow us to exit prematurely from the parser, put all of the state
into the `State.t` record.  The API of the parser now has a simple `run`
function that will run the parser on a given output, processing through
a user-specified state.  The internal `parse_to` will run the parser
until a stop condition is reached, and return.

The stop condition is not implemented yet, just the API.