weave: Generalize test framework
Create a more general delta test framework and rework the sccs testing
to use it.  This framework will be generally useful for testing our own
weave writing as that is implemented.

It is important to note that this removes a test that was testing
partial walks through the push parser.  This doesn't really belong in
the sccs test, and this code will get heavily exercised by the code
writing deltas.