weave: Implement delta writer
This implements the core of the delta writer.  This is a fairly complex
piece of code, and although it passes the tests, could use some cleanup
to make it more understandable.

This also only writes stub data for the various fields in the header.
These should be filled in with time stamps and meaningful tags passed
in.