Implement the signoff command
The signoff command takes two revisions (currently the second most
recent and the most recent) and compares the, showing what has changed.
Some simple benchmarks show this to be noticeably faster than the Rust
implementation, probably because this is implemented as a proper pull
parser, rather than the Rust version which is a push parser through a
channel.