Improve support for ocaml 4.12
Ocaml 4.12 widens the cases where warning 16 (parameter cannot be
erased) occurs.  Specifically, the function `with_new_temp` does not
have any unlabeled arguments after the optional argument.  For now, fix
this by adding a unit after the arguments.  It makes it a little more
clumsy to call, and a better fix might just be to make the compression
argument non-optional.